+++ spec/spec_helper.rb
@@ -1,11 +1,12 @@
 require 'bundler/setup'
 require 'super'
+require 'super/struct'
+
+Bundler.require(:development)
 
 RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'
 
-  # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
 
   config.expect_with :rspec do |c|