+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - unreleased
+### Added:
+- Super Struct Attribute decoder support.
+- Default Time Codec (Super Struct auto-decodes ISO 8601 strings).
+
 ## [0.2.1] - 2019-06-21
 ### Fixed:
 - Application.loader